requirements.txt
requests
pyttsx3
SpeechRecognition
wikipedia
pywhatkit
python-dotenv
pyautogui
elevenlabs
opencv-python
numpy
google-generativeai
edge-tts
sounddevice
soundfile
mediapipe
pyaudio